LayoutTests/http/tests/webgpu/webgpu/webgpu/shader/execution/expression/call/builtin/acos-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: OperationError: map async was not successful
CONSOLE MESSAGE: Unhandled Promise Rejection: OperationError: map async was not successful
CONSOLE MESSAGE: Unhandled Promise Rejection: OperationError: map async was not successful
CONSOLE MESSAGE: Unhandled Promise Rejection: OperationError: map async was not successful
CONSOLE MESSAGE: Unhandled Promise Rejection: OperationError: map async was not successful
CONSOLE MESSAGE: Unhandled Promise Rejection: OperationError: map async was not successful
CONSOLE MESSAGE: Unhandled Promise Rejection: OperationError: map async was not successful
CONSOLE MESSAGE: Unhandled Promise Rejection: OperationError: map async was not successful
CONSOLE MESSAGE: Unhandled Promise Rejection: OperationError: map async was not successful

Harness Error (FAIL), message = Unhandled rejection: map async was not successful

FAIL :abstract_float:inputSource="const";vectorize="_undef_" assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
      at (elided: only 2 shown)
 Reached unreachable code
FAIL :abstract_float:inputSource="const";vectorize=2 assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
      at (elided: only 2 shown)
 Reached unreachable code
FAIL :abstract_float:inputSource="const";vectorize=3 assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
    assert@http://127.0.0.1:8000/webgpu/common/util/util.js:37:20
    release@http://127.0.0.1:8000/webgpu/webgpu/util/device_pool.js:106:13
 Reached unreachable code
PASS :abstract_float:inputSource="const";vectorize=4
FAIL :f32:inputSource="const";vectorize="_undef_" assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
      at (elided: only 2 shown)
 Reached unreachable code
FAIL :f32:inputSource="const";vectorize=2 assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
      at (elided: only 2 shown)
 Reached unreachable code
FAIL :f32:inputSource="const";vectorize=3 assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
      at (elided: only 2 shown)
 Reached unreachable code
FAIL :f32:inputSource="const";vectorize=4 assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
      at (elided: only 2 shown)
 Reached unreachable code
FAIL :f32:inputSource="uniform";vectorize="_undef_" assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
      at (elided: only 2 shown)
 Reached unreachable code
PASS :f32:inputSource="uniform";vectorize=2
PASS :f32:inputSource="uniform";vectorize=3
PASS :f32:inputSource="uniform";vectorize=4
PASS :f32:inputSource="storage_r";vectorize="_undef_"
PASS :f32:inputSource="storage_r";vectorize=2
PASS :f32:inputSource="storage_r";vectorize=3
PASS :f32:inputSource="storage_r";vectorize=4
PASS :f32:inputSource="storage_rw";vectorize="_undef_"
PASS :f32:inputSource="storage_rw";vectorize=2
PASS :f32:inputSource="storage_rw";vectorize=3
PASS :f32:inputSource="storage_rw";vectorize=4
FAIL :f16:inputSource="const";vectorize="_undef_" assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
      at (elided: only 2 shown)
 Reached unreachable code
FAIL :f16:inputSource="const";vectorize=2 assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
      at (elided: only 2 shown)
 Reached unreachable code
FAIL :f16:inputSource="const";vectorize=3 assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
      at (elided: only 2 shown)
 Reached unreachable code
FAIL :f16:inputSource="const";vectorize=4 assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
      at (elided: only 2 shown)
 Reached unreachable code
FAIL :f16:inputSource="uniform";vectorize="_undef_" assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
      at (elided: only 2 shown)
 Reached unreachable code
PASS :f16:inputSource="uniform";vectorize=2
PASS :f16:inputSource="uniform";vectorize=3
PASS :f16:inputSource="uniform";vectorize=4
PASS :f16:inputSource="storage_r";vectorize="_undef_"
PASS :f16:inputSource="storage_r";vectorize=2
PASS :f16:inputSource="storage_r";vectorize=3
PASS :f16:inputSource="storage_r";vectorize=4
PASS :f16:inputSource="storage_rw";vectorize="_undef_"
PASS :f16:inputSource="storage_rw";vectorize=2
PASS :f16:inputSource="storage_rw";vectorize=3
PASS :f16:inputSource="storage_rw";vectorize=4
